Fix UpdateNode hashing for building solver expressions
UpdateNodes in ArrayExprHash were not related to arrays.
Due to the lifecycle of update nodes, this is a problem if a new UpdateNode
has the same pointer as a former one, which might be assigned to another expression.

This leads to using compleatly different solver expressions
resulting in wrong expressions.

Fix uses array pointer in addition to UpdateNode pointer.
One assumption here is, that array pointers don't change.
(Which is currently the case.)