Make killing of states deterministic
As we allocate memory for states only using the deterministic allocator
(if we use the deterministic mode), we know the size of allocated memory
for all the states.

We kill states (in case there is not enough memory) based on
deterministic size. This allows us to kill the states deterministically
based on the amount of memory used by the states only.

There is a drawback of this method:
1. We currently do not track external objects (only happens during
   initialization).
2. We do not track the memory used by KLEE itself. This can be tracked
   throught statistics.