Fix generation of initial values using CEX cache
Assignments used by the CEX cache used to imply concrete values (0)
for free symbolics during evaluation of expressions.
This has the side effect that if a conclussion contains symbolic
variables, which are independent of the premisse, the evaluation of the
conclusion can lead to either true or false.

Specifically with `computeValidity()` and `computeTruth()`, such
assignments can be added to the CEX cache. As the whole query is
used as a key in such case, the conclusion is indistinguishable
from the premiss.

This is an undesired behavior for generating initial values: If the path
constraints resemble a subset of constraints which match a previously
inserted query, despite an assignment, the concretization might lead to
an invalid assignment for the initial values.

To counteract that behavior, this patch forces the decision on how
to evaluate an assignment (concrete/symbolic) to time of evaluation
not time of creation of the assignment.

Depending on the invoked method of the CEX solver, a concretization is
possible and reduces the number of solver invocations (e.g.
`computeValidity()`, `computeTruth()`) or the concretization is avoided
and a solver is called.